requirements.txt
# Core
pyspark==3.5.1
kafka-python==2.0.2
confluent-kafka==2.3.0
fastavro==1.9.4
pyarrow==15.0.0

# Storage
pyiceberg[s3]==0.6.1
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.3
redis[hiredis]==5.0.3
psycopg2-binary==2.9.9

# NLP
transformers==4.39.3
torch==2.2.2
sentencepiece==0.2.0

# dbt
dbt-core==1.7.9
dbt-snowflake==1.7.3
dbt-utils==1.1.1

# Orchestration
apache-airflow==2.8.1
apache-airflow-providers-snowflake==5.3.0
apache-airflow-providers-slack==8.5.1

# Quality
great-expectations==0.18.12
openlineage-airflow==1.10.0
openlineage-python==1.10.0

# API
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.0

# Data generation (producers)
faker==24.9.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0
